+++ tests/test_model_class.py
@@ -167,6 +167,31 @@ def test_model_clip(tmpdir, example_wflow_model, clipped_wflow_model):
     _compare_wflow_models(clipped_wflow_model, mod1)
 
 
+def test_model_inverse_clip(tmpdir, example_wflow_model):
+    # Clip method options
+    region = {
+        "subbasin": [12.3006, 46.4324],
+        "wflow_streamorder": 4,
+    }
+
+    # Clip workflow, based on example model
+    example_wflow_model.read()
+    # Get number of active pixels from full model
+    n_pixels_full = example_wflow_model.grid["wflow_subcatch"].sum()
+    example_wflow_model.clip_grid(region, inverse_clip=True)
+    # Get number of active pixels from inversely clipped model
+    n_pixels_inverse_clipped = example_wflow_model.grid["wflow_subcatch"].sum()
+
+    # Do clipping again, but normally
+    example_wflow_model.read()
+    example_wflow_model.clip_grid(region, inverse_clip=False)
+    # Get number of active pixels from clipped model
+    n_pixels_clipped = example_wflow_model.grid["wflow_subcatch"].sum()
+
+    assert n_pixels_inverse_clipped < n_pixels_full
+    assert n_pixels_full == n_pixels_inverse_clipped + n_pixels_clipped
+
+
 def test_model_results(example_wflow_results):
     # Tests on results
     # Number of dict keys = 1 for output + 1 for netcdf + nb of csv.column